rename <Name>Slot.svelte to <Name>Snippet.svelte (more terminology change needed in docs)
- restore running end-to-end tests in GitHub Actions
- MultiSelect forward ...rest props to the input element
