+++ src/site/index.ts
@@ -1,7 +1,7 @@
-export { default as ColorSlot } from './ColorSlot.svelte'
+export { default as ColorSnippet } from './ColorSnippet.svelte'
 export { default as DemoNav } from './DemoNav.svelte'
 export { default as Examples } from './Examples.svx'
 export { default as Footer } from './Footer.svelte'
-export { default as LanguageSlot } from './LanguageSlot.svelte'
+export { default as LanguageSnippet } from './LanguageSnippet.svelte'
 export { default as MinusIcon } from './MinusIcon.svelte'
-export { default as RepoSlot } from './RepoSlot.svelte'
+export { default as RepoSnippet } from './RepoSnippet.svelte'
